fix: preserve bot catalog core boundary
