+++ src/backend/src/agentclaw/community/di/modules/bot_public_module.py
@@ -22,10 +22,7 @@ from __future__ import annotations
 from injector import Binder, Module, inject, provider, singleton
 
 from agentclaw.community.api.bot_discover_service import BotDiscoverServiceProtocol
-from agentclaw.community.api.bot_public_service import (
-    BotCatalogMetadataServiceProtocol,
-    BotPublicServiceProtocol,
-)
+from agentclaw.community.api.bot_public_service import BotPublicServiceProtocol
 from agentclaw.community.plugin_api.approval_workflow import ApprovalWorkflowPlugin
 from agentclaw.community.core.repository.protocols.bot import BotRepository
 from agentclaw.community.core.bot_management.services.bot_service import BotService
@@ -34,6 +31,9 @@ from agentclaw.community.core.bot_public.services.bot_discover_service import Bo
 from agentclaw.community.core.bot_public.services.bot_public_service import BotPublicService
 from agentclaw.community.core.bot_public.services.bot_catalog_metadata_service import (
     UnavailableBotCatalogMetadataService,
+)
+from agentclaw.community.core.bot_public.catalog_metadata import (
+    BotCatalogMetadataServiceProtocol,
 )
 from agentclaw.community.core.devices.services.device_context_resolver import (
     DeviceContextResolver,
